requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# HuggingFace ecosystem
transformers>=4.35.0
datasets>=2.14.0
evaluate>=0.4.0
tokenizers>=0.15.0

# Machine learning
scikit-learn>=1.3.0
scipy>=1.11.0

# PyTorch Lightning
pytorch-lightning>=2.0.0

# Parameter Efficient Fine-tuning
peft>=0.6.0

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Logging and experiment tracking
wandb>=0.15.0

# Utilities
tqdm>=4.65.0
